refactored to use browserify
