+++ .gitignore
@@ -1,8 +1,15 @@
 .*
+
+*.mp3
+*.m4a
+*.ogg
+*.wav
+*.mv4
+*.mp4
+*.webm
+
 node_modules/
-demo/audio
-*demo*/audio
-*test*/audio
+*/audio
 bower_components/
 /_*
