+++ helios-audio-mixer.js
@@ -1,1442 +1,2 @@
-/* global bowser, TWEEN */
-
-'use strict';
-
-var heliosAudioMixer = (function() {
-
-  // ######  ###### ###### ######  ##### ######
-  // ##   ## ##       ##   ##     ##       ##
-  // ##   ## #####    ##   #####  ##       ##
-  // ##   ## ##       ##   ##     ##       ##
-  // ######  ######   ##   ######  #####   ##
-
-  // Feature Detection
-
-  var Detect = {
-
-    // Web audio API support
-    webAudio: !!(window.AudioContext || window.webkitAudioContext),
-
-    // Which audio types can the browser actually play?
-    audioTypes: (function() {
-      var el = document.createElement('audio')
-
-      return {
-        '.m4a': !!(el.canPlayType && el.canPlayType('audio/mp4; codecs="mp4a.40.2"').replace(/no/, '')),
-        '.mp3': !!(el.canPlayType && el.canPlayType('audio/mpeg;').replace(/no/, '')),
-        '.ogg': !!(el.canPlayType && el.canPlayType('audio/ogg; codecs="vorbis"').replace(/no/, ''))
-      }
-    })(),
-
-    videoTypes: (function() {
-
-      var el = document.createElement('video')
-
-      return {
-        '.webm': !!(el.canPlayType && el.canPlayType('video/webm; codecs="vp8, vorbis"').replace(/no/, '')),
-        '.mp4':  !!(el.canPlayType && el.canPlayType('video/mp4; codecs="avc1.42E01E, mp4a.40.2"').replace(/no/, '')),
-        '.ogv':  !!(el.canPlayType && el.canPlayType('video/ogg; codecs="theora"').replace(/no/, ''))
-      }
-    })(),
-
-    // prefer bowser, but with fallback
-    browser: (function() {
-
-      if(typeof bowser !== 'undefined') return bowser;
-
-      return {
-        firefox: !!navigator.userAgent.match(/Firefox/g),
-        android: !!navigator.userAgent.match(/Android/g),
-        msie:    !!navigator.userAgent.match(/MSIE/g),
-        ios:     !!navigator.userAgent.match(/(iPad|iPhone|iPod)/g),
-
-        version: false
-      };
-    })(),
-
-    // is tween.js present?
-    tween: (function() {
-      return (typeof TWEEN === 'undefined') ? false : true
-    })()
-
-  }
-
-  // ######   #####   ####  ######
-  // ##   ## ##   ## ##     ##
-  // ######  #######  ####  #####
-  // ##   ## ##   ##     ## ##
-  // ######  ##   ## #####  ######
-
-  // shared functionality: event system, extend
-
-  var BaseClass = function() {};
-
-  BaseClass.prototype.on = function(type, callback) {
-    this.events[type] = this.events[type] || [];
-    this.events[type].push(callback);
-
-    return this
-  };
-
-  BaseClass.prototype.one = function(type, callback) {
-    var _this = this
-    this.events[type] = this.events[type] || [];
-    this.events[type].push(function(){
-      _this.off(type)
-      callback()
-    });
-
-    return this
-  };
-
-  BaseClass.prototype.off = function(type) {
-    if(type === '*')
-      this.events = {};
-    else
-      this.events[type] = [];
-
-    return this
-  };
-
-  BaseClass.prototype.trigger = function(type) {
-
-    if(!this.events[type]) return;
-
-    var args = Array.prototype.slice.call(arguments, 1);
-
-    for (var i = 0, l = this.events[type].length; i < l; i++)
-        if(typeof this.events[type][i] === 'function')
-          this.events[type][i].apply(this, args);
-
-  };
-
-  BaseClass.prototype.extend = function() {
-    var output = {}
-    var args = arguments
-    var l = args.length
-
-    for (var i = 0; i < l; i++)
-      for (var key in args[i])
-        if(args[i].hasOwnProperty(key))
-          output[key] = args[i][key];
-    return output;
-  };
-
-  var extend = function() {
-    var output = {}
-    var args = arguments
-    var l = args.length
-
-    for (var i = 0; i < l; i++)
-      for (var key in args[i])
-        if(args[i].hasOwnProperty(key))
-          output[key] = args[i][key];
-    return output;
-  };
-
-
-
-
-
-
-
-
-
-
-
-
-  var constrain = function(val, min, max) {
-    if(val < min) return min;
-    if(val > max) return max;
-    return val;
-  };
-
-
-
-
-
-
-
-
-  // ###  ### #### ##   ##
-  // ########  ##   ## ##
-  // ## ## ##  ##    ###
-  // ##    ##  ##   ## ##
-  // ##    ## #### ##   ##
-
-  var Mix = function(opts) {
-
-    this.Track = Track
-
-    this.debug = 1 // 0 no logging, 1 minimal, 2 all (spammy)
-
-    this.log = function(lvl) {
-      if(lvl <= this.debug) {
-        var str = ''
-        for (var i = 1; i < arguments.length; i++)
-          str += arguments[i] + ' '
-        console.log(str)
-      }
-    }
-
-    this.setLogLvl = function(lvl) {
-      this.debug = constrain(lvl, 0, 2);
-      this.log(1, '[Mixer] Set log level:', lvl)
-    }
-
-    var defaults = {
-      fileTypes: [ '.mp3', '.m4a', '.ogg' ],
-      html5: false,
-      gain: 1, // master gain for entire mix
-    }
-    this.options = extend(defaults, opts || {})
-
-
-    this.tracks  = [];    // tracks as numbered array
-    this.lookup  = {};    // tracks as lookup table: lookup['trackname']
-
-    this.muted   = false; // master mute status
-
-    this.context = null;  // AudioContext object (if webAudio is available)
-
-    this.detect  = Detect; // external reference to Detect object
-
-
-    // Web Audio support overrides
-    // ********************************************************
-
-    if(
-      (Detect.browser.name === 'Firefox' && Detect.version && Detect.version < 25) || // Firefox < 25
-      (Detect.browser.ios === true && Detect.browser.version === 7) ||                 // ios 7
-      this.options.html5
-   ) {
-      Detect.webAudio = false;
-    }
-
-
-    // File Types
-    // ********************************************************
-
-    for (var i = this.options.fileTypes.length - 1; i >= 0; i--) {
-      if(!Detect.audioTypes[ this.options.fileTypes[i] ])
-        this.options.fileTypes.splice(i, 1);
-    }
-
-    if(this.options.fileTypes.length <= 0) {
-      console.warn('[Mixer] Can’t initialize: none of the specified audio types can play in this browser.')
-      return;
-    }
-
-
-    // Initialize
-    // ********************************************************
-
-    if(Detect.webAudio === true) {
-      this.context = (typeof AudioContext === 'function' ? new window.AudioContext() : new window.webkitAudioContext() )
-    }
-
-    this.log(1, '[Mixer] initialized,', (Detect.webAudio ? 'Web Audio Mode,' : 'HTML5 Mode,'), 'can play:', this.options.fileTypes)
-
-    return this
-
-  };
-
-  // inherit base class functionality
-  Mix.prototype = new BaseClass()
-
-
-  /**************************************************************************
-
-    Track Management
-
-  **************************************************************************/
-
-  Mix.prototype.createTrack = function(name, opts) {
-
-    var _this = this;
-    var track;
-
-    if(!name)
-      _this.log(0, '[Mixer] Can’t create track with no name')
-
-    if(Detect.webAudio === true) {
-
-      if(this.lookup[name]) {
-        _this.log(0, '[Mixer] a track named “' + _this.name + '” already exists')
-        return
-      }
-
-
-      track = new Track(name, opts, _this);
-
-      this.tracks.push(track);
-      this.lookup[name] = track;
-
-    } else {
-
-      track = this.lookup[name];
-
-      if(!track) { // create new track
-
-        track = new Track(name, opts, _this);
-
-        this.tracks.push(track);
-        this.lookup[ name ] = track;
-
-      } else { // change source
-
-        track.pause();
-
-        track.options = track.extend(track, track.defaults, opts || {});
-
-        if(track.options.source) {
-          track.options.source  += _this.mix.options.fileTypes[0];
-          track.createHTML5elementSource();
-        }
-      }
-    }
-
-    return track;
-
-  };
-
-
-  Mix.prototype.removeTrack = function(_input) {
-
-    var _this = this
-
-    var track, name
-
-    if(typeof _input === 'string')
-      name = _input
-    else if(typeof _input === 'object' && _input.name)
-      name = _input.name
-
-    var track = _this.lookup[name]
-
-    if(!track) {
-      _this.log(1, '[Mixer] can’t remove "' + name + '", it doesn’t exist');
-      return;
-    }
-
-    if(Detect.webAudio === true) {
-
-      track.pause();
-
-      var rest
-      var arr   = _this.tracks
-      var total = arr.length
-
-      for (var i = 0; i < total; i++) {
-        if(arr[i] && arr[i].name === name) {
-          rest = arr.slice(i + 1 || total);
-          arr.length = (i < 0) ? (total + i) : (i);
-          arr.push.apply(arr, rest);
-        }
-      }
-
-      track.trigger('remove', _this);
-      track.events = [];
-
-      // stop memory leaks!
-      if(track.options.sourceMode === 'element' && track.element)
-        track.element.src = '';
-
-      track = null;
-      delete _this.lookup[name];
-      _this.log(1, '[Mixer] Removed track "' + name + '"');
-
-    } else {
-
-      track.pause();
-
-      track.element.src = '';
-      track.source = '';
-
-      track.ready = false;
-      track.loaded = false;
-
-    }
-
-  };
-
-
-  Mix.prototype.getTrack = function(name) {
-    return this.lookup[name] || false;
-  };
-
-
-
-
-
-  /**************************************************************************
-
-    Global Mix Control
-
-  **************************************************************************/
-
-  Mix.prototype.pause = function() {
-
-    this.log(1, '[Mixer] Pausing ' + this.tracks.length + ' track(s) ||')
-
-    for (var i = 0; i < this.tracks.length; i++)
-      this.tracks[i].pause()
-  };
-
-  Mix.prototype.play = function() {
-
-    this.log(1, '[Mixer] Playing ' + this.tracks.length + ' track(s) >')
-
-    for (var i = 0; i < this.tracks.length; i++)
-      this.tracks[i].play()
-  };
-
-  Mix.prototype.stop = function() {
-
-    this.log(1, '[Mixer] Stopping ' + this.tracks.length + ' track(s) .')
-
-    for (var i = 0; i < this.tracks.length; i++)
-       this.tracks[i].stop()
-  };
-
-
-
-
-  Mix.prototype.mute = function() {
-    if(this.muted) return
-    this.muted = true
-    this.log(1, '[Mixer] Muting ' + this.tracks.length + ' tracks')
-    for (var i = 0; i < this.tracks.length; i++)
-      this.tracks[i].mute();
-  };
-
-
-  Mix.prototype.unmute = function() {
-    if(!this.muted) return
-    this.muted = false
-    this.log(1, '[Mixer] Unmuting ' + this.tracks.length + ' tracks')
-    for (var i = 0; i < this.tracks.length; i++)
-      this.tracks[i].unmute();
-  };
-
-
-
-  Mix.prototype.gain = function(masterGain) {
-    if(typeof masterGain !== 'undefined') {
-      this.options.gain = masterGain;
-
-      // this seems silly, but tracks multiply their gain by the master's
-      // so if we change the master gain and call track.gain() we will
-      // get the intended result
-      for (var i = 0; i < this.tracks.length; i++)
-        this.tracks[i].gain(this.tracks[i].gain());
-    }
-
-    return this.options.gain;
-  }
-
-
-
-
-  /**************************************************************************
-
-    Utilities
-
-  **************************************************************************/
-
-
-  // call this using requestanimationframe
-  Mix.prototype.updateTween = function() {
-    TWEEN.update();
-  };
-
-
-
-
-  Mix.prototype.report = function(){
-    var log = ""
-    for (var i = 0; i < this.tracks.length; i++) {
-      log += this.tracks[i].gain() + '\t' + this.tracks[i].currentTime() + '\t' + this.tracks[i].name + '\n'
-    }
-    console.log(log)
-  }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-  // ###### #####   #####   ##### ##  ##
-  //   ##   ##  ## ##   ## ##     ## ##
-  //   ##   #####  ####### ##     ####
-  //   ##   ##  ## ##   ## ##     ## ##
-  //   ##   ##  ## ##   ##  ##### ##  ##
-
-
-  var Track = function(name, opts, mix) {
-
-    this.defaults = {
-
-      sourceMode: 'buffer', // buffer or (media) element
-
-      source: false,   // either path to audio source (without file extension) or b) html5 <audio> or <video> element
-
-      nodes:      [],  // array of strings: names of desired additional audio nodes
-
-      gain:        1,  // initial/current gain (0-1)
-      gainCache:   false,  // for resuming from mute
-
-      pan:         0,  // circular horizontal pan
-
-      panX:        0,  // real 3d pan
-      panY:        0,  //
-      panZ:        0,  //
-
-      start:       0,  // start time in seconds
-      cachedTime:  0,  // local current time (cached for resuming from pause)
-      startTime:   0,  // time started (cached for accurately reporting currentTime)
-
-      looping:  false, //
-      autoplay: true,  // play immediately on load
-      muted:    (mix.muted) ? true : false
-    };
-
-    // override option defaults
-    this.options = this.extend.call(this, this.defaults, opts || {});
-
-    if(this.options.gainCache === false)
-      this.options.gainCache = this.options.gain
-
-    this.name = name;
-
-    // Status
-    this.status = {
-      loaded:  false,
-      ready:   false,
-      playing: false
-    }
-
-    this.mix     = mix;  // reference to parent
-
-    this.events  = {};
-    this.tweens  = {};
-    this.nodes   = undefined;   // holds the web audio nodes (gain and pan are defaults, all other optional)
-
-    this.onendtimer = undefined;  // web audio api in chrome doesn't support onend event yet :(
-
-    this.element = undefined; // html5 <audio> or <video> element
-    this.source  = undefined; //  web audio source:
-
-    if(typeof this.options.source === 'string' && this.options.source.indexOf('blob:') !== 0) {
-      // append extension only if it’s a file path
-      this.options.source  += this.mix.options.fileTypes[0];
-    } else if(typeof this.options.source === 'object') {
-      // otherwise switch mode to element
-      this.options.sourceMode = 'element'
-    }
-
-    this.mix.log(1, '[Mixer] createTrack "' + this.name + '", mode: "' + this.options.sourceMode + '", autoplay: ' + this.options.autoplay);
-
-
-
-    // Load
-    // ~~~~
-
-    if(Detect.webAudio === true) {
-
-      // Web Audio
-
-      if(!this.options.source) {
-        this.mix.log(1, '[Mixer] Creating track "' + name + '" without a source');
-        return;
-      }
-
-      if(this.options.sourceMode === 'buffer') {
-
-        this.loadBufferSource()
-
-      } else if(this.options.sourceMode === 'element') {
-
-        if(typeof this.options.source === 'object')
-          this.useHTML5elementSource()
-        else
-          this.createHTML5elementSource()
-
-      }
-
-    } else {
-
-      // HTML5
-
-      this.mix.log(1, '[Mixer] creating html5 element for track ' + name);
-
-      // Look for pre-created audio element and failing that create one
-      this.element = document.querySelector('audio#' + name);
-
-      if(!this.element) {
-        var el = document.createElement('audio');
-        el.id = name;
-        document.body.appendChild(el);
-        this.element = document.querySelector('audio#' + name);
-      }
-
-      var canplay = function() {
-        if(this.status.loaded) return;
-
-        this.status.loaded = true;
-        this.status.ready = true;
-
-        if(!this.options.autoplay)
-          this.pause();
-
-        this.trigger('load', this);
-      }
-
-      var _this = this
-
-      // canplaythrough listener
-      _this.element.addEventListener('canplaythrough', canplay, false);
-      _this.element.addEventListener('load',           canplay, false);
-
-      // Looping
-      _this.element.addEventListener('ended', function() {
-
-        if(_this.options.looping) {
-
-          _this.mix.log(2, 'Track ' + _this.name + ' looping')
-
-          _this.element.currentTime = 0;
-          _this.element.play();
-
-        } else {
-
-          _this.trigger('ended', _this);
-          _this.mix.removeTrack(_this.name);
-        }
-
-      }, false);
-
-      this.createHTML5elementSource(this.options.source);
-    }
-  };
-
-
-
-  Track.prototype = new BaseClass()
-
-
-
-
-  // ##    ######   #####  ######
-  // ##   ##    ## ##   ## ##   ##
-  // ##   ##    ## ####### ##   ##
-  // ##   ##    ## ##   ## ##   ##
-  // ##### ######  ##   ## ######
-
-  // Create out-of-DOM html5 <audio> element as source
-  Track.prototype.createHTML5elementSource = function() {
-
-    var _this = this;
-    if(!_this.options.source) return;
-
-    _this.mix.log(2, '[Mixer] Track "' + this.name + '" creating HTML5 element source: "' + _this.options.source + _this.mix.options.fileTypes[0]  + '"');
-    _this.status.ready = false
-
-    var src = _this.options.source
-
-    _this.options.source = document.createElement('audio')
-    _this.options.source.crossOrigin = ''
-    _this.options.source.src = src
-
-    _this.useHTML5elementSource()
-  }
-
-  // Use existing in-DOM html5 <audio> or <video> element as source
-  Track.prototype.useHTML5elementSource = function() {
-
-    var _this = this;
-    if(!_this.options.source) return;
-
-    _this.mix.log(2, '[Mixer] Track "' + this.name + '" use HTML5 element source: "' + _this.options.source + '"')
-
-    _this.element = _this.options.source
-    _this.options.source.crossOrigin = ''
-    _this.options.source = _this.element.src
-
-    /**
-     * Add options if they're set.
-     */
-    if (_this.options.looping) { _this.element.loop  = true; }
-    if (_this.options.muted)   { _this.element.muted = true; }
-
-    if(_this.mix.context)
-      _this.source = _this.mix.context.createMediaElementSource(_this.element);
-
-    var ready = function() {
-      _this.status.loaded = true
-      if(_this.options.autoplay) _this.play();
-      _this.trigger('load', _this);
-    }
-
-    _this.element.addEventListener('canplaythrough', ready)
-    _this.element.addEventListener('error', function() { _this.trigger('loadError') })
-
-    _this.element.load()
-
-    return _this
-  }
-
-  Track.prototype.loadBufferSource = function(forcePlay) {
-
-    var _this = this;
-    if(!_this.options.source) return;
-
-    this.mix.log(2, '[Mixer] Track "' + this.name + '" webAudio source: "' + _this.options.source + '"')
-
-    var request = new XMLHttpRequest();
-    request.open('GET', _this.options.source, true);
-    request.responseType = 'arraybuffer';
-
-    // asynchronous callback
-    request.onload = function() {
-      _this.mix.log(2, '[Mixer] "' + _this.name + '" loaded "' + _this.options.source + '"');
-      _this.options.audioData = request.response; // cache the audio data
-      _this.status.loaded = true;
-      _this.trigger('load', _this);
-      if(forcePlay){
-        _this.play(true)
-      } else {
-        if(_this.options.autoplay) _this.play();
-      }
-    }
-
-    request.onerror = function() {
-      _this.mix.log(1, '[Mixer] couldn’t load track "' + _this.name + '" with source "' + _this.options.source + '"')
-      _this.trigger('loadError', _this)
-    }
-
-    request.send();
-  }
-
-
-
-
-
-
-
-
-
-  // ###  ##  ######  ######  ######  ####
-  // #### ## ##    ## ##   ## ##     ##
-  // ## #### ##    ## ##   ## #####   ####
-  // ##  ### ##    ## ##   ## ##         ##
-  // ##   ##  ######  ######  ###### #####
-
-
-  // AudioNode Creation
-  // -> this function can be chained
-
-  Track.prototype.addNode = function(nodeType) {
-
-    var _this = this;
-
-    // if this is the first time we’re calling addNode, connect directly to the source
-    if(!_this.nodes.lastnode) _this.nodes.lastnode = _this.source;
-
-    this.mix.log(2, ' +  addNode ' + nodeType);
-
-    // Analyser and Processor ********************************************************
-
-    if(nodeType === 'analyse') {
-
-      // create a script processor with bufferSize of 2048
-      _this.nodes.processor = _this.mix.context.createScriptProcessor(2048, 1, 1)
-
-      // create an analyser
-      _this.nodes.analyser = _this.mix.context.createAnalyser()
-      _this.nodes.analyser.smoothingTimeConstant = 0.5
-      _this.nodes.analyser.fftSize = 32
-
-      _this.nodes.processor.connect(_this.mix.context.destination) // processor -> destination
-      _this.nodes.analyser.connect(_this.nodes.processor)          // analyser -> processor
-
-      // define a Uint8Array to receive the analyser’s data
-      _this.options.bufferLength = new Uint8Array(_this.nodes.analyser.frequencyBinCount)
-      var analyserData = new Uint8Array(_this.options.bufferLength)
-
-      _this.nodes.lastnode.connect(_this.nodes.analyser)
-
-      _this.nodes.processor.onaudioprocess = function(){
-        _this.nodes.analyser.getByteTimeDomainData(analyserData)
-        _this.trigger('analyse', analyserData)
-      }
-    }
-
-    // Gain ********************************************************
-    // http://www.w3.org/TR/webaudio/#GainNode
-
-    else if(nodeType === 'gain') {
-
-      if(_this.mix.context.createGainNode)
-        _this.nodes.gain = _this.mix.context.createGainNode();
-      else
-        _this.nodes.gain = _this.mix.context.createGain();
-
-      // connect last created node to newly created node
-      _this.nodes.lastnode.connect(_this.nodes.gain);
-
-      // set newly created node to last node in the chain
-      _this.nodes.lastnode = _this.nodes.gain;
-
-    }
-
-    // Panner ********************************************************
-    // http://www.w3.org/TR/webaudio/#PannerNode
-
-    else if(nodeType === 'panner') {
-
-      if(window.AudioContext) {
-
-        _this.nodes.panner = _this.mix.context.createPanner();
-        // _this.nodes.panner.panningModel = 'equalpower';
-        // _this.nodes.panner.panningModel = 'HRTF';
-
-      } else if(window.webkitAudioContext) {
-
-        _this.nodes.panner = _this.mix.context.createPanner();
-        // _this.nodes.panner.panningModel = webkitAudioPannerNode.EQUALPOWER;
-        // _this.nodes.panner.panningModel = _this.nodes.panner.EQUALPOWER;
-
-      } else {
-        return _this;
-      }
-
-      _this.nodes.lastnode.connect(_this.nodes.panner);
-      _this.nodes.lastnode = _this.nodes.panner;
-
-    }
-
-    // Convolver (Reverb etc) **********************************************
-    // http://www.w3.org/TR/webaudio/#ConvolverNode
-
-    else if(nodeType === 'convolver') {
-
-      if(!_this.mix.context.createConvolver) return _this;
-
-      _this.nodes.convolver = _this.mix.context.createConvolver();
-
-      // TODO: implement loading impulse response for the convolver node
-      // http://chromium.googlecode.com/svn/trunk/samples/audio/impulse-responses/
-
-      // _this.nodes.convolver.buffer = convolverBuffer;
-
-
-    }
-
-    // Compressor ********************************************************
-    // http://www.w3.org/TR/webaudio/#DynamicsCompressorNode
-
-    else if(nodeType === 'compressor') {
-
-      _this.nodes.compressor = _this.mix.context.createDynamicsCompressor();
-
-      // no settings required really…
-
-      _this.nodes.lastnode.connect(_this.nodes.compressor);
-      _this.nodes.lastnode = _this.nodes.compressor;
-
-    }
-
-    // Delay ********************************************************
-    // http://www.w3.org/TR/webaudio/#DelayNode
-
-    else if(nodeType === 'delay') {
-
-      if(Detect.nodes.delayNode)
-        _this.nodes.delay = _this.mix.context.createDelayNode();
-      else
-        _this.nodes.delay = _this.mix.context.createDelay();
-
-      _this.nodes.delay.delayTime = 0;
-
-      _this.nodes.lastnode.connect(_this.nodes.delay);
-      _this.nodes.lastnode = _this.nodes.delay;
-
-    }
-
-    // it’s chainable
-    return _this;
-  }
-
-
-  // ######  ##     ##### ##    ##
-  // ##   ## ##    ##   ## ##  ##
-  // ######  ##    #######  ####
-  // ##      ##    ##   ##   ##
-  // ##      ##### ##   ##   ##
-
-  Track.prototype.play = function(bufferSourceLoaded) {
-
-    var _this = this;
-
-    if(!_this.status.loaded) {
-      this.mix.log(1, 'Can’t play track "' + _this.name + '", not loaded')
-      return;
-    }
-
-    if(_this.status.playing === true) return;
-    _this.status.playing = true;
-
-    if(!Detect.webAudio)
-      playSingleElement(_this)
-
-    else if(Detect.webAudio && _this.options.sourceMode === 'buffer'){
-      // need to re-xhr the audio file so we loop back to load
-      if(bufferSourceLoaded)
-        playBufferSource(_this)
-      else {
-        _this.status.playing = false;
-        _this.loadBufferSource(true) // loop back to load
-      }
-    }
-
-    else if(Detect.webAudio && _this.options.sourceMode === 'element')
-      playElementSource(_this)
-
-
-    return _this
-  }
-
-
-  function playCreateNodes(_this) {
-
-    _this.mix.log(2, 'Creating nodes for track "' + _this.name + '"')
-
-    // Create Nodes
-    // ~~~~~~~~~~~~
-
-    _this.nodes = {}
-
-    // 1. Create standard nodes (gain and pan)
-    _this.addNode('panner').addNode('gain')
-
-    // 2. Create additional nodes
-    if(_this.options.nodes.length){
-      for (var i = 0; i < _this.options.nodes.length; i++) {
-        _this.addNode( _this.options.nodes[i] );
-      }
-    }
-
-    // 3. Connect the last node in the chain to the destination
-    _this.nodes.lastnode.connect(_this.mix.context.destination);
-
-  }
-
-  // ********************************************************
-
-  function playElementSource(_this) {
-
-    // unlike buffer mode, we only need to construct the nodes once
-    if( !_this.nodes ) {
-
-      playCreateNodes(_this)
-
-      // we also only want one event listener
-      _this.element.addEventListener('ended', function() {
-        _this.trigger('ended', _this)
-      }, false)
-    }
-
-
-    // Apply Options
-    // ~~~~~~~~~~~~~~
-
-    _this.status.ready = true;
-    _this.trigger('ready', _this);
-
-    if(_this.options.looping) _this.element.loop = true;
-    else                      _this.element.loop = false;
-
-    _this.gain(_this.options.gain)
-    _this.pan(_this.options.pan)
-
-    // Start Time
-
-    _this.options.startTime = _this.element.currentTime - _this.options.cachedTime;
-    var startFrom = _this.options.cachedTime || 0;
-
-    _this.mix.log(1, '[Mixer] Playing track "' + _this.name + '" from ' + startFrom + ' (' + _this.options.startTime + ') gain ' + _this.gain());
-
-    // Play!
-
-    _this.element.currentTime = startFrom;
-    _this.element.play()
-
-    _this.trigger('play', _this)
-
-  }
-
-
-  // ********************************************************
-
-  function setEndTimer(){
-    var _this = this
-    var startFrom = _this.options.cachedTime || 0
-    var timerDuration = (_this.source.buffer.duration - startFrom)
-
-    _this.onendtimer = setTimeout(function() {
-      _this.trigger('ended', _this)
-
-      if(_this.options.looping){
-
-        if(bowser && bowser.chrome && Math.floor(bowser.version) >= 42){
-          // HACK chrome 42+ looping fix
-          _this.stop()
-          _this.play()
-        } else {
-          setEndTimer.call(_this)
-        }
-
-      }
-
-    }, timerDuration * 1000)
-  }
-
-  function playBufferSource(_this) {
-
-    _this.status.ready = false
-
-    // Construct Audio Buffer
-    // ~~~~~~~~~~~~~~~~~~~~~~
-
-    // (we have to re-construct the buffer every time we begin play)
-
-    _this.source = null
-
-    // *sigh* async makes for such messy code
-    var finish = function() {
-
-      playCreateNodes(_this)
-
-      _this.status.ready = true;
-      _this.trigger('ready', _this);
-
-      // Apply Options
-      _this.source.loop = (_this.options.looping) ? true : false
-      _this.gain(_this.options.gain)
-      _this.pan(_this.options.pan)
-
-
-      // Play
-      // ~~~~
-
-      _this.options.startTime = _this.source.context.currentTime - _this.options.cachedTime;
-      var startFrom = _this.options.cachedTime || 0;
-
-      _this.mix.log(2, '[Mixer] Playing track "' + _this.name + '" from ' + startFrom + ' (' + _this.options.startTime + ') gain ' + _this.gain());
-
-      // prefer start() but fall back to deprecated noteOn()
-      if(typeof _this.source.start === 'function') _this.source.start(0, startFrom);
-      else                                         _this.source.noteOn(startFrom);
-
-      // fake ended event
-      _this.onendtimer = false
-      setEndTimer.call(_this)
-
-
-      _this.trigger('play', _this)
-
-    }
-
-
-    // Create source
-    // ~~~~~~~~~~~~~
-
-    // W3C standard implementation (Firefox, recent Chrome)
-    if(typeof _this.mix.context.createGain === 'function') {
-
-      _this.mix.context.decodeAudioData(_this.options.audioData, function(decodedBuffer){
-        if(_this.status.ready) return
-
-        _this.source        = _this.mix.context.createBufferSource();
-        var sourceBuffer    = decodedBuffer;
-        _this.source.buffer = sourceBuffer;
-
-        finish()
-      })
-    }
-
-    // Non-standard Webkit implementation (Safari, old Chrome)
-    else if(typeof _this.mix.context.createGainNode === 'function') {
-
-      _this.source = _this.mix.context.createBufferSource();
-      var sourceBuffer  = _this.mix.context.createBuffer(_this.options.audioData, true);
-      _this.source.buffer = sourceBuffer;
-
-      finish()
-    }
-  }
-
-  // ********************************************************
-
-  function playSingleElement(_this) {
-
-    _this.mix.log(1, '[Mixer] Playing track "' + _this.name + '" >')
-
-    _this.gain(_this.options.gain)
-
-    _this.status.ready  = true;
-    _this.element.play();
-    _this.trigger('play', _this);
-  }
-
-
-  // ######   #####  ##   ##  ####  ######
-  // ##   ## ##   ## ##   ## ##     ##
-  // ######  ####### ##   ##  ####  #####
-  // ##      ##   ## ##   ##     ## ##
-  // ##      ##   ##  #####  #####  ######
-
-
-  Track.prototype.pause = function(at) {
-
-    if(!this.status.ready || !this.status.playing) return;
-
-    var _this = this;
-
-    // cache time to resume from later
-    if(typeof at === 'number') _this.options.cachedTime = at;
-    else                         _this.options.cachedTime = _this.currentTime();
-
-    _this.status.playing = false;
-
-    if(_this.onendtimer) clearTimeout(_this.onendtimer);
-
-    if(Detect.webAudio === true) {
-
-      if(_this.options.sourceMode === 'buffer') {
-
-        // prefer stop(), fallback to deprecated noteOff()
-        if(typeof _this.source.stop === 'function')
-          _this.source.stop(0);
-        else if(typeof _this.source.noteOff === 'function')
-          _this.source.noteOff(0);
-
-      } else if(_this.options.sourceMode === 'element') {
-
-        _this.element.pause()
-      }
-
-    } else {
-
-      _this.element.pause();
-    }
-
-    _this.mix.log(2, '[Mixer] Pausing track "' + _this.name + '" at ' + _this.options.cachedTime)
-    _this.trigger('pause', _this);
-
-    return _this
-  }
-
-
-  //  #### ###### ######  ######
-  // ##      ##  ##    ## ##   ##
-  //  ####   ##  ##    ## ######
-  //     ##  ##  ##    ## ##
-  // #####   ##   ######  ##
-
-  Track.prototype.stop = function() {
-
-    if(!this.status.ready || !this.status.playing) return;
-
-    var _this = this;
-
-    if(!_this.status.playing) return
-
-    if(_this.onendtimer) clearTimeout(_this.onendtimer)
-
-    _this.options.cachedTime = _this.options.startTime = 0
-
-    if(Detect.webAudio === true && _this.options.sourceMode === 'buffer') {
-
-      // prefer stop(), fallback to deprecated noteOff()
-      if(typeof _this.source.stop === 'function')
-        _this.source.stop(0);
-      else if(typeof _this.source.noteOff === 'function')
-        _this.source.noteOff(0);
-
-    } else {
-
-      _this.options.autoplay = false;
-
-      _this.element.pause();
-      _this.element.currentTime = 0;
-    }
-
-    _this.mix.log(2, '[Mixer] Stopping track "' + _this.name)
-    _this.status.playing = false
-    _this.trigger('stop', _this);
-
-    _this.options.gain = _this.options.gainCache;
-
-    return _this
-
-  }
-
-
-  // ######   #####  ###  ##
-  // ##   ## ##   ## #### ##
-  // ######  ####### ## ####
-  // ##      ##   ## ##  ###
-  // ##      ##   ## ##   ##
-
-  // proper 3d stereo panning
-  Track.prototype.pan = function(angleDeg) {
-
-    if( !Detect.webAudio || !this.status.ready || !this.nodes.panner ) return
-
-    if(typeof angleDeg === 'string') {
-      if(     angleDeg === 'front') angleDeg =   0;
-      else if(angleDeg === 'back' ) angleDeg = 180;
-      else if(angleDeg === 'left' ) angleDeg = 270;
-      else if(angleDeg === 'right') angleDeg =  90;
-    }
-
-    if(typeof angleDeg === 'number') {
-
-      this.options.pan = angleDeg % 360;
-
-      var angleRad = (-angleDeg + 90) * 0.017453292519943295; // * PI/180
-
-      var x = this.options.panX = Math.cos(angleRad);
-      var y = this.options.panY = Math.sin(angleRad);
-      var z = this.options.panZ = -0.5;
-
-      this.nodes.panner.setPosition(x, y, z);
-
-      this.trigger('pan', this.options.pan, this)
-
-      return this // all setters should be chainable
-    }
-
-    return this.options.pan
-  }
-
-  Track.prototype.tweenPan = function(angleDeg, tweenDuration) {
-    var _this = this;
-
-    return new Promise(function(resolve, reject){
-      if( !Detect.tween || !Detect.webAudio || !_this.status.ready || !_this.nodes.panner ) reject(Error('nope nope nope'))
-      if( typeof angleDeg !== 'number' || typeof tweenDuration !== 'number' ) reject(Error('Not a valid tween duration.'))
-
-      _this.mix.log(2, '[Mixer] "' + _this.name + '" tweening pan2d')
-
-      if(_this.tweens.pan) _this.tweens.pan.stop()
-
-      _this.tweens.pan = new TWEEN.Tween({ currentAngle: _this.options.pan })
-        .to({ currentAngle: angleDeg }, tweenDuration)
-        .easing(TWEEN.Easing.Sinusoidal.InOut)
-        .onUpdate(function() {
-          _this.pan(this.currentAngle)
-        })
-        .onComplete(function() {
-          resolve(_this)
-        })
-        .start()
-
-      return _this
-
-    })
-  }
-
-
-  //  #####   #####  #### ###  ##
-  // ##      ##   ##  ##  #### ##
-  // ##  ### #######  ##  ## ####
-  // ##   ## ##   ##  ##  ##  ###
-  //  #####  ##   ## #### ##   ##
-
-  // cache current gain for restoring later
-  Track.prototype.gainCache = function(setTo) {
-    if(typeof setTo === 'number') {
-      this.options.gainCache = setTo;
-      return this
-    } else {
-      return this.options.gainCache
-    }
-  }
-
-  // gain range 0-1
-  Track.prototype.gain = function(val) {
-    if(typeof val === 'number') {
-
-      val = constrain(val, 0, 1) // normalize value
-
-      if(this.options.muted) {
-        this.gainCache(val) // cache the value
-        this.options.gain = 0
-      } else {
-        this.options.gain = val
-      }
-
-      if(this.status.playing) {
-
-        if(!Detect.webAudio)
-          this.element.volume = this.options.gain * this.mix.options.gain
-        else if( this.nodes && this.nodes.gain )
-          this.nodes.gain.gain.value = this.options.gain * this.mix.options.gain
-
-      }
-
-      this.mix.log(2, '[Mixer] "' + this.name + '" setting gain to ' + this.options.gain)
-
-      this.trigger('gain', this.options.gain, this)
-
-      // setters should be chainable
-      return this
-
-    }
-
-    return this.options.gain
-
-  }
-
-  Track.prototype.tweenGain = function(_val, _tweenDuration) {
-    var _this = this;
-    return new Promise(function(resolve, reject){
-      if(typeof _val !== 'number' || typeof _tweenDuration !== 'number') reject(Error('Invalid value for duration.'))
-      _this.mix.log(2, '[Mixer] "' + _this.name + '" tweening gain ' + _this.options.gain + ' -> ' + _val + ' ('+_tweenDuration+'ms)')
-
-      // replace existing gain tween
-      if(_this.tweens.gain) _this.tweens.gain.stop()
-
-      _this.tweens.gain = new TWEEN.Tween({ currentGain: _this.options.gain })
-        .to({ currentGain: _val }, _tweenDuration)
-        .easing(TWEEN.Easing.Sinusoidal.InOut)
-        .onUpdate(function() {
-          _this.gain(this.currentGain)
-        })
-        .onComplete(function() {
-          resolve(_this)
-        })
-        .start()
-
-    })
-  }
-
-  Track.prototype.mute = function() {
-    this.gainCache(this.options.gain)
-    this.gain(0)
-    this.options.muted = true
-    if(this.options.sourceMode === 'element')
-      this.element.muted = true
-    return this
-  }
-
-  Track.prototype.unmute = function() {
-    this.options.muted = false;
-    if(this.options.sourceMode === 'element')
-      this.element.muted = false
-    this.gain(this.options.gainCache)
-    return this
-  }
-
-
-  //  ##### ##   ## #####  ###### ###  ## ###### ###### #### ###  ### ######
-  // ##     ##   ## ##  ## ##     #### ##   ##     ##    ##  ######## ##
-  // ##     ##   ## #####  #####  ## ####   ##     ##    ##  ## ## ## #####
-  // ##     ##   ## ##  ## ##     ##  ###   ##     ##    ##  ##    ## ##
-  //  #####  #####  ##  ## ###### ##   ##   ##     ##   #### ##    ## ######
-
-  Track.prototype.currentTime = function(setTo) {
-    if(!this.status.ready) return;
-
-    if(typeof setTo === 'number') {
-
-      this.mix.log(2, '[Mixer] setting track "' + this.name + '" to time', setTo)
-
-      if(Detect.webAudio && this.options.sourceMode === 'buffer') {
-
-        if(this.status.playing) {
-          this.pause(setTo);
-          this.play();
-        } else {
-          this.options.cachedTime = setTo;
-        }
-
-      } else {
-
-        this.element.currentTime = setTo;
-      }
-
-      return this
-    }
-
-    if(!this.status.playing) return this.options.cachedTime || 0;
-
-    if(Detect.webAudio && this.options.sourceMode === 'buffer')
-      return this.source.context.currentTime - this.options.startTime || 0;
-    else
-      return this.element.currentTime || 0;
-  }
-
-  /**************************************************************************
-
-    Getters Only
-
-  **************************************************************************/
-
-  function timeFormat(seconds) {
-    var m = Math.floor(seconds / 60) < 10 ? '0' + Math.floor(seconds / 60) : Math.floor(seconds / 60);
-    var s = Math.floor(seconds - (m * 60)) < 10 ? '0' + Math.floor(seconds - (m * 60)) : Math.floor(seconds - (m * 60));
-    return m + ':' + s;
-  }
-
-  Track.prototype.formattedTime = function(includeDuration) {
-    if(!this.status.ready) return;
-
-    if(includeDuration)
-      return timeFormat(this.currentTime()) + '/' + timeFormat(this.duration());
-    else
-      return timeFormat(this.currentTime());
-  }
-
-  Track.prototype.duration = function() {
-    if(!this.status.ready) return;
-
-    if(Detect.webAudio && this.options.sourceMode === 'buffer')
-      return this.source.buffer.duration || 0;
-    else
-      return this.element.duration || 0;
-  }
-
-  return Mix
-
-}());
+!function e(t,o,n){function i(r,a){if(!o[r]){if(!t[r]){var u="function"==typeof require&&require;if(!a&&u)return u(r,!0);if(s)return s(r,!0);var c=new Error("Cannot find module '"+r+"'");throw c.code="MODULE_NOT_FOUND",c}var p=o[r]={exports:{}};t[r][0].call(p.exports,function(e){var o=t[r][1][e];return i(o?o:e)},p,p.exports,e,t,o,n)}return o[r].exports}for(var s="function"==typeof require&&require,r=0;r<n.length;r++)i(n[r]);return i}({1:[function(e,t,o){var n=e("./mix");window.heliosAudioMixer=n},{"./mix":4}],2:[function(e,t,o){var n=e("./utils"),i={};t.exports=i,i.level=1,i.log=function(e){if(e<=i.level){for(var t="[Mixer] ",o=1;o<arguments.length;o++)t+=arguments[o]+" ";console.log(t)}},i.setLogLvl=function(e){this.debug=n.constrain(e,0,2),i.log(1,"Set log level:",e)}},{"./utils":6}],3:[function(e,t,o){var n={};n.webAudio=!(!window.AudioContext&&!window.webkitAudioContext),n.audioTypes=function(){var e=document.createElement("audio");return{".m4a":!(!e.canPlayType||!e.canPlayType('audio/mp4; codecs="mp4a.40.2"').replace(/no/,"")),".mp3":!(!e.canPlayType||!e.canPlayType("audio/mpeg;").replace(/no/,"")),".ogg":!(!e.canPlayType||!e.canPlayType('audio/ogg; codecs="vorbis"').replace(/no/,""))}}(),n.videoTypes=function(){var e=document.createElement("video");return{".webm":!(!e.canPlayType||!e.canPlayType('video/webm; codecs="vp8, vorbis"').replace(/no/,"")),".mp4":!(!e.canPlayType||!e.canPlayType('video/mp4; codecs="avc1.42E01E, mp4a.40.2"').replace(/no/,"")),".ogv":!(!e.canPlayType||!e.canPlayType('video/ogg; codecs="theora"').replace(/no/,""))}}(),n.browser=function(){return"undefined"!=typeof bowser?bowser:{firefox:!!navigator.userAgent.match(/Firefox/g),android:!!navigator.userAgent.match(/Android/g),msie:!!navigator.userAgent.match(/MSIE/g),ios:!!navigator.userAgent.match(/(iPad|iPhone|iPod)/g),version:!1}}(),n.tween=function(){return"undefined"==typeof TWEEN?!1:!0}(),t.exports=n},{}],4:[function(e,t,o){var n=e("./utils"),i=e("./track"),s=e("./detect"),r=e("./debug"),a=function(e){this.Track=i,this.setLogLvl=r.setLogLvl;var t={fileTypes:[".mp3",".m4a",".ogg"],html5:!1,gain:1};this.options=n.extend(t,e||{}),this.tracks=[],this.lookup={},this.muted=!1,this.context=null,this.detect=s,("Firefox"===s.browser.name&&s.version&&s.version<25||s.browser.ios===!0&&7===s.browser.version||this.options.html5)&&(s.webAudio=!1);for(var o=this.options.fileTypes.length-1;o>=0;o--)s.audioTypes[this.options.fileTypes[o]]||this.options.fileTypes.splice(o,1);return this.options.fileTypes.length<=0?void console.warn("Can’t initialize: none of the specified audio types can play in this browser."):(s.webAudio===!0&&(this.context="function"==typeof AudioContext?new window.AudioContext:new window.webkitAudioContext),r.log(1,"initialized,",s.webAudio?"Web Audio Mode,":"HTML5 Mode,","can play:",this.options.fileTypes),this)};a.prototype.on=n.events.on,a.prototype.one=n.events.one,a.prototype.off=n.events.off,a.prototype.trigger=n.events.trigger,a.prototype.createTrack=function(e,t){var o,a=this;if(e||r.log(0,"Can’t create track with no name"),s.webAudio===!0){if(this.lookup[e])return void r.log(0,"a track named “"+a.name+"” already exists");o=new i(e,t,a),this.tracks.push(o),this.lookup[e]=o}else o=this.lookup[e],o?(o.pause(),o.options=n.extend(o,o.defaults,t||{}),o.options.source&&(o.options.source+=a.mix.options.fileTypes[0],o.createHTML5elementSource())):(o=new i(e,t,a),this.tracks.push(o),this.lookup[e]=o);return o},a.prototype.removeTrack=function(e){var t,o,n=this;"string"==typeof e?o=e:"object"==typeof e&&e.name&&(o=e.name);var t=n.lookup[o];if(!t)return void r.log(1,'can’t remove "'+o+'", it doesn’t exist');if(s.webAudio===!0){t.pause();for(var i,a=n.tracks,u=a.length,c=0;u>c;c++)a[c]&&a[c].name===o&&(i=a.slice(c+1||u),a.length=0>c?u+c:c,a.push.apply(a,i));t.trigger("remove",n),t.events=[],"element"===t.options.sourceMode&&t.element&&(t.element.src=""),t=null,delete n.lookup[o],r.log(1,'Removed track "'+o+'"')}else t.pause(),t.element.src="",t.source="",t.ready=!1,t.loaded=!1},a.prototype.getTrack=function(e){return this.lookup[e]||!1},a.prototype.pause=function(){r.log(1,"Pausing "+this.tracks.length+" track(s) ||");for(var e=0;e<this.tracks.length;e++)this.tracks[e].pause()},a.prototype.play=function(){r.log(1,"Playing "+this.tracks.length+" track(s) >");for(var e=0;e<this.tracks.length;e++)this.tracks[e].play()},a.prototype.stop=function(){r.log(1,"Stopping "+this.tracks.length+" track(s) .");for(var e=0;e<this.tracks.length;e++)this.tracks[e].stop()},a.prototype.mute=function(){if(!this.muted){this.muted=!0,r.log(1,"Muting "+this.tracks.length+" tracks");for(var e=0;e<this.tracks.length;e++)this.tracks[e].mute()}},a.prototype.unmute=function(){if(this.muted){this.muted=!1,r.log(1,"Unmuting "+this.tracks.length+" tracks");for(var e=0;e<this.tracks.length;e++)this.tracks[e].unmute()}},a.prototype.gain=function(e){if("undefined"!=typeof e){this.options.gain=e;for(var t=0;t<this.tracks.length;t++)this.tracks[t].gain(this.tracks[t].gain())}return this.options.gain},a.prototype.updateTween=function(){TWEEN.update()},a.prototype.report=function(){for(var e="",t=0;t<this.tracks.length;t++)e+=this.tracks[t].gain()+"	"+this.tracks[t].currentTime()+"	"+this.tracks[t].name+"\n";console.log(e)},t.exports=a},{"./debug":2,"./detect":3,"./track":5,"./utils":6}],5:[function(e,t,o){function n(e){if(p.log(2,'Creating nodes for track "'+e.name+'"'),e.nodes={},e.addNode("panner").addNode("gain"),e.options.nodes.length)for(var t=0;t<e.options.nodes.length;t++)e.addNode(e.options.nodes[t]);e.nodes.lastnode.connect(e.mix.context.destination)}function i(e){e.nodes||(n(e),e.element.addEventListener("ended",function(){e.trigger("ended",e)},!1)),e.status.ready=!0,e.trigger("ready",e),e.options.looping?e.element.loop=!0:e.element.loop=!1,e.gain(e.options.gain),e.pan(e.options.pan),e.options.startTime=e.element.currentTime-e.options.cachedTime;var t=e.options.cachedTime||0;p.log(1,'Playing track "'+e.name+'" from '+t+" ("+e.options.startTime+") gain "+e.gain()),e.element.currentTime=t,e.element.play(),e.trigger("play",e)}function s(){var e=this,t=e.options.cachedTime||0,o=e.source.buffer.duration-t;e.onendtimer=setTimeout(function(){e.trigger("ended",e),e.options.looping&&(bowser&&bowser.chrome&&Math.floor(bowser.version)>=42?(e.stop(),e.play()):s.call(e))},1e3*o)}function r(e){e.status.ready=!1,e.source=null;var t=function(){n(e),e.status.ready=!0,e.trigger("ready",e),e.source.loop=e.options.looping?!0:!1,e.gain(e.options.gain),e.pan(e.options.pan),e.options.startTime=e.source.context.currentTime-e.options.cachedTime;var t=e.options.cachedTime||0;p.log(2,'Playing track "'+e.name+'" from '+t+" ("+e.options.startTime+") gain "+e.gain()),"function"==typeof e.source.start?e.source.start(0,t):e.source.noteOn(t),e.onendtimer=!1,s.call(e),e.trigger("play",e)};if("function"==typeof e.mix.context.createGain)e.mix.context.decodeAudioData(e.options.audioData,function(o){if(!e.status.ready){e.source=e.mix.context.createBufferSource();var n=o;e.source.buffer=n,t()}});else if("function"==typeof e.mix.context.createGainNode){e.source=e.mix.context.createBufferSource();var o=e.mix.context.createBuffer(e.options.audioData,!0);e.source.buffer=o,t()}}function a(e){p.log(1,'Playing track "'+e.name+'" >'),e.gain(e.options.gain),e.status.ready=!0,e.element.play(),e.trigger("play",e)}function u(e){var t=Math.floor(e/60)<10?"0"+Math.floor(e/60):Math.floor(e/60),o=Math.floor(e-60*t)<10?"0"+Math.floor(e-60*t):Math.floor(e-60*t);return t+":"+o}var c=e("./utils"),p=(e("./detect"),e("./debug")),d=function(e,t,o){if(this.defaults={sourceMode:"buffer",source:!1,nodes:[],gain:1,gainCache:!1,pan:0,panX:0,panY:0,panZ:0,start:0,cachedTime:0,startTime:0,looping:!1,autoplay:!0,muted:o.muted?!0:!1},this.options=c.extend(this.defaults,t||{}),this.options.gainCache===!1&&(this.options.gainCache=this.options.gain),this.name=e,this.status={loaded:!1,ready:!1,playing:!1},this.mix=o,this.events={},this.tweens={},this.nodes=void 0,this.onendtimer=void 0,this.element=void 0,this.source=void 0,"string"==typeof this.options.source&&0!==this.options.source.indexOf("blob:")?this.options.source+=this.mix.options.fileTypes[0]:"object"==typeof this.options.source&&(this.options.sourceMode="element"),p.log(1,'createTrack "'+this.name+'", mode: "'+this.options.sourceMode+'", autoplay: '+this.options.autoplay),Detect.webAudio===!0){if(!this.options.source)return void p.log(1,'Creating track "'+e+'" without a source');"buffer"===this.options.sourceMode?this.loadBufferSource():"element"===this.options.sourceMode&&("object"==typeof this.options.source?this.useHTML5elementSource():this.createHTML5elementSource())}else{if(p.log(1,"creating html5 element for track "+e),this.element=document.querySelector("audio#"+e),!this.element){var n=document.createElement("audio");n.id=e,document.body.appendChild(n),this.element=document.querySelector("audio#"+e)}var i=function(){this.status.loaded||(this.status.loaded=!0,this.status.ready=!0,this.options.autoplay||this.pause(),this.trigger("load",this))},s=this;s.element.addEventListener("canplaythrough",i,!1),s.element.addEventListener("load",i,!1),s.element.addEventListener("ended",function(){s.options.looping?(p.log(2,"Track "+s.name+" looping"),s.element.currentTime=0,s.element.play()):(s.trigger("ended",s),s.mix.removeTrack(s.name))},!1),this.createHTML5elementSource(this.options.source)}};d.prototype.on=c.events.on,d.prototype.one=c.events.one,d.prototype.off=c.events.off,d.prototype.trigger=c.events.trigger,d.prototype.createHTML5elementSource=function(){var e=this;if(e.options.source){p.log(2,'Track "'+this.name+'" creating HTML5 element source: "'+e.options.source+e.mix.options.fileTypes[0]+'"'),e.status.ready=!1;var t=e.options.source;e.options.source=document.createElement("audio"),e.options.source.crossOrigin="",e.options.source.src=t,e.useHTML5elementSource()}},d.prototype.useHTML5elementSource=function(){var e=this;if(e.options.source){p.log(2,'Track "'+this.name+'" use HTML5 element source: "'+e.options.source+'"'),e.element=e.options.source,e.options.source.crossOrigin="",e.options.source=e.element.src,e.options.looping&&(e.element.loop=!0),e.options.muted&&(e.element.muted=!0),e.mix.context&&(e.source=e.mix.context.createMediaElementSource(e.element));var t=function(){e.status.loaded=!0,e.options.autoplay&&e.play(),e.trigger("load",e)};return e.element.addEventListener("canplaythrough",t),e.element.addEventListener("error",function(){e.trigger("loadError")}),e.element.load(),e}},d.prototype.loadBufferSource=function(e){var t=this;if(t.options.source){p.log(2,'Track "'+this.name+'" webAudio source: "'+t.options.source+'"');var o=new XMLHttpRequest;o.open("GET",t.options.source,!0),o.responseType="arraybuffer",o.onload=function(){p.log(2,'"'+t.name+'" loaded "'+t.options.source+'"'),t.options.audioData=o.response,t.status.loaded=!0,t.trigger("load",t),e?t.play(!0):t.options.autoplay&&t.play()},o.onerror=function(){p.log(1,'couldn’t load track "'+t.name+'" with source "'+t.options.source+'"'),t.trigger("loadError",t)},o.send()}},d.prototype.addNode=function(e){var t=this;if(t.nodes.lastnode||(t.nodes.lastnode=t.source),p.log(2," +  addNode "+e),"analyse"===e){t.nodes.processor=t.mix.context.createScriptProcessor(2048,1,1),t.nodes.analyser=t.mix.context.createAnalyser(),t.nodes.analyser.smoothingTimeConstant=.5,t.nodes.analyser.fftSize=32,t.nodes.processor.connect(t.mix.context.destination),t.nodes.analyser.connect(t.nodes.processor),t.options.bufferLength=new Uint8Array(t.nodes.analyser.frequencyBinCount);var o=new Uint8Array(t.options.bufferLength);t.nodes.lastnode.connect(t.nodes.analyser),t.nodes.processor.onaudioprocess=function(){t.nodes.analyser.getByteTimeDomainData(o),t.trigger("analyse",o)}}else if("gain"===e)t.mix.context.createGainNode?t.nodes.gain=t.mix.context.createGainNode():t.nodes.gain=t.mix.context.createGain(),t.nodes.lastnode.connect(t.nodes.gain),t.nodes.lastnode=t.nodes.gain;else if("panner"===e){if(window.AudioContext)t.nodes.panner=t.mix.context.createPanner();else{if(!window.webkitAudioContext)return t;t.nodes.panner=t.mix.context.createPanner()}t.nodes.lastnode.connect(t.nodes.panner),t.nodes.lastnode=t.nodes.panner}else if("convolver"===e){if(!t.mix.context.createConvolver)return t;t.nodes.convolver=t.mix.context.createConvolver()}else"compressor"===e?(t.nodes.compressor=t.mix.context.createDynamicsCompressor(),t.nodes.lastnode.connect(t.nodes.compressor),t.nodes.lastnode=t.nodes.compressor):"delay"===e&&(Detect.nodes.delayNode?t.nodes.delay=t.mix.context.createDelayNode():t.nodes.delay=t.mix.context.createDelay(),t.nodes.delay.delayTime=0,t.nodes.lastnode.connect(t.nodes.delay),t.nodes.lastnode=t.nodes.delay);return t},d.prototype.play=function(e){var t=this;if(!t.status.loaded)return void p.log(1,'Can’t play track "'+t.name+'", not loaded');if(t.status.playing!==!0)return t.status.playing=!0,Detect.webAudio?Detect.webAudio&&"buffer"===t.options.sourceMode?e?r(t):(t.status.playing=!1,t.loadBufferSource(!0)):Detect.webAudio&&"element"===t.options.sourceMode&&i(t):a(t),t},d.prototype.pause=function(e){if(this.status.ready&&this.status.playing){var t=this;return"number"==typeof e?t.options.cachedTime=e:t.options.cachedTime=t.currentTime(),t.status.playing=!1,t.onendtimer&&clearTimeout(t.onendtimer),Detect.webAudio===!0?"buffer"===t.options.sourceMode?"function"==typeof t.source.stop?t.source.stop(0):"function"==typeof t.source.noteOff&&t.source.noteOff(0):"element"===t.options.sourceMode&&t.element.pause():t.element.pause(),p.log(2,'Pausing track "'+t.name+'" at '+t.options.cachedTime),t.trigger("pause",t),t}},d.prototype.stop=function(){if(this.status.ready&&this.status.playing){var e=this;if(e.status.playing)return e.onendtimer&&clearTimeout(e.onendtimer),e.options.cachedTime=e.options.startTime=0,Detect.webAudio===!0&&"buffer"===e.options.sourceMode?"function"==typeof e.source.stop?e.source.stop(0):"function"==typeof e.source.noteOff&&e.source.noteOff(0):(e.options.autoplay=!1,e.element.pause(),e.element.currentTime=0),p.log(2,'Stopping track "'+e.name),e.status.playing=!1,e.trigger("stop",e),e.options.gain=e.options.gainCache,e}},d.prototype.pan=function(e){if(Detect.webAudio&&this.status.ready&&this.nodes.panner){if("string"==typeof e&&("front"===e?e=0:"back"===e?e=180:"left"===e?e=270:"right"===e&&(e=90)),"number"==typeof e){this.options.pan=e%360;var t=.017453292519943295*(-e+90),o=this.options.panX=Math.cos(t),n=this.options.panY=Math.sin(t),i=this.options.panZ=-.5;return this.nodes.panner.setPosition(o,n,i),this.trigger("pan",this.options.pan,this),this}return this.options.pan}},d.prototype.tweenPan=function(e,t){var o=this;return new Promise(function(n,i){return Detect.tween&&Detect.webAudio&&o.status.ready&&o.nodes.panner||i(Error("nope nope nope")),("number"!=typeof e||"number"!=typeof t)&&i(Error("Not a valid tween duration.")),p.log(2,'"'+o.name+'" tweening pan2d'),o.tweens.pan&&o.tweens.pan.stop(),o.tweens.pan=new TWEEN.Tween({currentAngle:o.options.pan}).to({currentAngle:e},t).easing(TWEEN.Easing.Sinusoidal.InOut).onUpdate(function(){o.pan(this.currentAngle)}).onComplete(function(){n(o)}).start(),o})},d.prototype.gainCache=function(e){return"number"==typeof e?(this.options.gainCache=e,this):this.options.gainCache},d.prototype.gain=function(e){return"number"==typeof e?(e=c.constrain(e,0,1),this.options.muted?(this.gainCache(e),this.options.gain=0):this.options.gain=e,this.status.playing&&(Detect.webAudio?this.nodes&&this.nodes.gain&&(this.nodes.gain.gain.value=this.options.gain*this.mix.options.gain):this.element.volume=this.options.gain*this.mix.options.gain),p.log(2,'"'+this.name+'" setting gain to '+this.options.gain),this.trigger("gain",this.options.gain,this),this):this.options.gain},d.prototype.tweenGain=function(e,t){var o=this;return new Promise(function(n,i){("number"!=typeof e||"number"!=typeof t)&&i(Error("Invalid value for duration.")),p.log(2,'"'+o.name+'" tweening gain '+o.options.gain+" -> "+e+" ("+t+"ms)"),o.tweens.gain&&o.tweens.gain.stop(),o.tweens.gain=new TWEEN.Tween({currentGain:o.options.gain}).to({currentGain:e},t).easing(TWEEN.Easing.Sinusoidal.InOut).onUpdate(function(){o.gain(this.currentGain)}).onComplete(function(){n(o)}).start()})},d.prototype.mute=function(){return this.gainCache(this.options.gain),this.gain(0),this.options.muted=!0,"element"===this.options.sourceMode&&(this.element.muted=!0),this},d.prototype.unmute=function(){return this.options.muted=!1,"element"===this.options.sourceMode&&(this.element.muted=!1),this.gain(this.options.gainCache),this},d.prototype.currentTime=function(e){return this.status.ready?"number"==typeof e?(p.log(2,'setting track "'+this.name+'" to time',e),Detect.webAudio&&"buffer"===this.options.sourceMode?this.status.playing?(this.pause(e),this.play()):this.options.cachedTime=e:this.element.currentTime=e,this):this.status.playing?Detect.webAudio&&"buffer"===this.options.sourceMode?this.source.context.currentTime-this.options.startTime||0:this.element.currentTime||0:this.options.cachedTime||0:void 0},d.prototype.formattedTime=function(e){return this.status.ready?e?u(this.currentTime())+"/"+u(this.duration()):u(this.currentTime()):void 0},d.prototype.duration=function(){return this.status.ready?Detect.webAudio&&"buffer"===this.options.sourceMode?this.source.buffer.duration||0:this.element.duration||0:void 0},t.exports=d},{"./debug":2,"./detect":3,"./utils":6}],6:[function(e,t,o){var n={};t.exports=n,n.extend=function(){for(var e={},t=arguments,o=t.length,n=0;o>n;n++)for(var i in t[n])t[n].hasOwnProperty(i)&&(e[i]=t[n][i]);return e},n.constrain=function(e,t,o){return t>e?t:e>o?o:e},n.events={},n.events.on=function(e,t){return this.events[e]=this.events[e]||[],this.events[e].push(t),this},n.events.one=function(e,t){var o=this;return this.events[e]=this.events[e]||[],this.events[e].push(function(){o.off(e),t()}),this},n.events.off=function(e){return"*"===e?this.events={}:this.events[e]=[],this},n.events.trigger=function(e){if(this.events[e])for(var t=Array.prototype.slice.call(arguments,1),o=0,n=this.events[e].length;n>o;o++)"function"==typeof this.events[e][o]&&this.events[e][o].apply(this,t)}},{}]},{},[1]);
+//# sourceMappingURL=helios-audio-mixer.js.map
