+++ src/bundle-for-browserify.js
@@ -1,0 +1,2 @@
+var mix = require('./mix');
+module.exports = mix;
