+++ src/bundle-standalone.js
@@ -1,0 +1,2 @@
+var mix = require('./mix');
+window.heliosAudioMixer = mix;
