--- test/index.html
@@ -1,37 +1,0 @@
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Mocha Tests</title>
-  <link rel="stylesheet" href="../bower_components/mocha/mocha.css" />
-</head>
-<body>
-
-  <!-- <div id="requirements">
-    <p>
-      This test requires 
-    </p>
-  </div> -->
-
-  <div id="mocha"></div>
-
-  <script src="../bower_components/chai/chai.js"></script>
-  <script src="../bower_components/mocha/mocha.js"></script>
-  <script>
-    mocha.setup('bdd')
-    var expect = chai.expect;
-
-  </script>
-
-  <script src="../bower_components/bowser/bowser.min.js"></script>
-  <!-- <script src="../bower_components/tweenjs/src/Tween.js"></script> -->
-  <script src="../bower_components/helios-frame-runner/helios-frame-runner.js"></script>
-  <script src="../helios-audio-mixer.js"></script>
-  <script src="test.js"></script>
-
-  <script>
-    mocha.checkLeaks();
-    mocha.run();
-  </script>
-
-</body>
-</html>
